requirements.txt
# Cryptocurrency Policy Analyzer Requirements

# Web Framework
flask==3.1.0
flask-cors==5.0.0

# Environment & Configuration
python-dotenv==1.0.1

# OpenAI
openai==1.68.2

# LangChain Components
langchain==0.3.21
langchain-core==0.3.47
langchain-openai==0.3.9
langchain-elasticsearch==0.3.2
langchain-text-splitters==0.3.7
langchain-experimental==0.3.4
langchain-community==0.3.20

# Database & Storage
elasticsearch==8.17.2

# Data Validation & Processing
pydantic==2.10.6
numpy==1.26.4
requests==2.32.3
python-docx==1.1.2
pypdf==4.3.1

# NLP & Embeddings
sentence-transformers==3.3.0

# Document Processing
docling==2.28.0

# Graph & Data Structures
networkx==3.4.2
pandas==2.2.3

# JSON Processing
json_repair==0.40.0

# LlamaIndex
llama-index==0.10.15
llama-index-embeddings-openai==0.1.11
llama-index-readers-docling==0.3.2 